pyproject.toml
[project]
name = "fit-track"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.24.0",
    "aiohttp>=3.13.3",
    "asyncio>=4.0.0",
    "python-dotenv>=1.2.1",
    "matplotlib>=3.8.0",
    "numpy>=1.26.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
